Move custom-roles view model to separate file
